\name{eddyD}
\alias{df}
\docType{data}
\title{
Surface Renewal 30 minute 20 Hz Data
}
\description{
This is a dataset from UNL -SNR with moisture (gm-3) records over 30 minutes
}
\usage{data("eddyD ")}
\format{
  A data frame with 36000 observations on the following 1 variable.
}
\details{
UNL
}
\source{
UNL
}
\references{
UNL
}
\keyword{datasets}
