% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddysrt.R
\name{eddysrt.R#1}
\alias{eddysrt}
\alias{eddysrt.R#1}
\title{: Surface Renewal Parameters}
\usage{
eddysrt(z, df)
}
\arguments{
\item{z}{20 rows derived from 20 Hz per second}

\item{df}{The dataframe consisting of  30 minute - 20 Hz file (30 * 20 * 60)=36000 values}
}
\value{
a numeric matrix vector
}
\description{
:  Calculating Surface Renewal Parameters for a 30 minute - 20 Hz Dataset

Read in the file

This function will take in the 30 minute -20 Hz dataset and calculate tau, amplitude and r
}
\details{
eddysrt

The matrix vector:rvalue (seconds) - maximum ratio of 3rd moment to r, tau (total ramp duration) and amplitude of ramp
}
\seealso{
{NISTunits}
}

